+++ lib/qonfig/command_set.rb
@@ -14,6 +14,7 @@ module Qonfig
     # @since 0.1.0
     def initialize
       @commands = []
+      @__access_lock__ = Mutex.new
     end
 
     # @param command [Qonfig::Commands::Base]
@@ -22,7 +23,7 @@ module Qonfig
     # @api private
     # @since 0.1.0
     def add_command(command)
-      commands << command
+      thread_safe { commands << command }
     end
     alias_method :<<, :add_command
 
@@ -41,7 +42,28 @@ module Qonfig
     # @api private
     # @since 0.1.0
     def concat(command_set)
-      commands.concat(command_set.commands)
+      thread_safe { commands.concat(command_set.commands) }
+    end
+
+    # @return [Qonfig::CommandSet]
+    #
+    # @api private
+    # @since 0.2.0
+    def dup
+      thread_safe do
+        self.class.new.tap { |duplicate| duplicate.concat(self) }
+      end
+    end
+
+    private
+
+    # @param block [Proc]
+    # @return [Object]
+    #
+    # @api private
+    # @since 0.2.0
+    def thread_safe(&block)
+      @__access_lock__.synchronize(&block)
     end
   end
 end